adding dir checks

+++ config_oh-my-zsh.sh
@@ -11,10 +11,12 @@ if [ ! -n "$ZSH" ]; then
   ZSH=~/.oh-my-zsh
 fi
 
-rm $ZSH/plugins/battery/battery.plugin.zsh
-wget -O $ZSH/plugins/battery/battery.plugin.zsh https://raw.githubusercontent.com/initcfg/bin/master/oh-my-zsh/plugins/battery/battery.plugin.zsh
+if [ -d "$ZSH" ]; then
+  rm $ZSH/plugins/battery/battery.plugin.zsh
+  wget -O $ZSH/plugins/battery/battery.plugin.zsh https://raw.githubusercontent.com/initcfg/bin/master/oh-my-zsh/plugins/battery/battery.plugin.zsh
 
-wget -O $ZSH/themes/initcfg.zsh-theme https://raw.githubusercontent.com/initcfg/bin/master/oh-my-zsh/themes/initcfg.zsh-theme
+  wget -O $ZSH/themes/initcfg.zsh-theme https://raw.githubusercontent.com/initcfg/bin/master/oh-my-zsh/themes/initcfg.zsh-theme
 
-git -C $ZSH add -A
-git -C $ZSH commit -m "initcfg updates"
+  git -C $ZSH add -A
+  git -C $ZSH commit -m "initcfg updates"
+fi
